Update Plan, Goal, and Planner instance creation, add methods for date and time conversion
This commit updates how instances of Plan, Goal, and Planner are created in plannerControl, goalControl, and planControl. Also, two new methods, ConvertPlanDate and ConvertPlanTime, are added to the Plan struct in the Plan model for date and time conversion. Specifically, this simplifies the process of instance creation and provides centralized date and time conversion.