Refactor plan model and add plan instance generator
Removed testing functions from the 'plan.go' file, restructured it to be a construction and utility file for the 'Plan' struct. A method for generating a plan instance was added, along with methods for converting date and time string inputs into usable time.Time values. Each field of the 'Plan' struct was also explicitly defined.