#FLOW-15 Add user prompt and handling for plan goalID
Updated the CLI application to prompt users for a goalID when creating a plan and added GoalId field to the Plan struct in the internal model. This change enables goal-oriented tracking and presentation in the plan management function. Also, modified the corresponding request handlers and generators to accommodate this newly added information.