FLOW-17
Refine goal and plan output, add input validation

The output formatting of goal details in CLI has been simplified for clarity, no longer including plans detail. Exceptional handling added in CLI to catch and notify when no commands are entered. In planControl, the response is built differently to ensure array of plan models is returned. A minor edit in goalControl was made for better readability.