Implement command handling in chat applications
Additional functionality was added to the chat loop of sv1.go and pv1.go to check if user input starts with a $, signaling a command. Currently, the only command that can be handled is "history", which retrieves and displays the chat history. The implementation paves the way for future commands to be easily added.

Signed-off-by: Ola Yeku <ooyeku@gmail.com>